requirements.txt
mysql-connector-python==8.0.32
requests==2.28.2
DateTime==5.0
termcolor==2.2.0
selenium==4.8.0
PyVirtualDisplay~=3.0
cloudscraper==1.2.68
beautifulsoup4==4.11.2
python-dotenv~=0.21.1